correct exception class casing and readme file
